+++ cubed/__init__.py
@@ -14,9 +14,7 @@ except Exception:  # pragma: no cover
 
 from .array_api import Array
 from .core.array import (
-    Callback,
     Spec,
-    TaskEndEvent,
     compute,
     measure_reserved_mem,
     measure_reserved_memory,
@@ -25,6 +23,7 @@ from .core.array import (
 from .core.gufunc import apply_gufunc
 from .core.ops import from_array, from_zarr, map_blocks, store, to_zarr
 from .nan_functions import nanmean, nansum
+from .runtime.types import Callback, TaskEndEvent
 
 __all__ = [
     "__version__",
