+++ cubed/extensions/history.py
@@ -4,8 +4,8 @@ from pathlib import Path
 
 import pandas as pd
 
-from cubed.core.array import Callback
 from cubed.core.plan import visit_nodes
+from cubed.runtime.types import Callback
 
 
 class HistoryCallback(Callback):
