+++ cubed/runtime/executors/beam.py
@@ -8,9 +8,8 @@ import apache_beam as beam
 import networkx as nx
 from apache_beam.runners.runner import PipelineState
 
-from cubed.core.array import TaskEndEvent
 from cubed.core.plan import visit_nodes
-from cubed.runtime.types import DagExecutor
+from cubed.runtime.types import DagExecutor, TaskEndEvent
 from cubed.vendor.rechunker.types import Config, NoArgumentStageFunction, Stage
 
 from ..utils import gensym
