+++ cubed/runtime/executors/coiled.py
@@ -3,9 +3,9 @@ from typing import Any, Mapping, Optional, Sequence
 import coiled
 from networkx import MultiDiGraph
 
-from cubed.core.array import Callback, Spec
+from cubed.core.array import Spec
 from cubed.core.plan import visit_nodes
-from cubed.runtime.types import DagExecutor
+from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import execution_stats, handle_callbacks
