+++ cubed/runtime/executors/dask_distributed_async.py
@@ -17,11 +17,11 @@ from aiostream.core import Stream
 from dask.distributed import Client
 from networkx import MultiDiGraph
 
-from cubed.core.array import Callback, Spec
+from cubed.core.array import Spec
 from cubed.core.plan import visit_node_generations, visit_nodes
 from cubed.primitive.types import CubedPipeline
 from cubed.runtime.executors.asyncio import async_map_unordered
-from cubed.runtime.types import DagExecutor
+from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import execution_stats, gensym, handle_callbacks
