+++ cubed/runtime/executors/python.py
@@ -2,10 +2,10 @@ from typing import Any, Callable, Optional, Sequence
 
 from networkx import MultiDiGraph
 
-from cubed.core.array import Callback, Spec, TaskEndEvent
+from cubed.core.array import Spec
 from cubed.core.plan import visit_nodes
 from cubed.primitive.types import CubedPipeline
-from cubed.runtime.types import DagExecutor
+from cubed.runtime.types import Callback, DagExecutor, TaskEndEvent
 
 
 def exec_stage_func(func: Callable[..., Any], *args, **kwargs):
